util: conditionally build CGO functions
Some projects don't need the CGO functions in the util package;
split out the functions into a separate file so CGO_ENABLED=0
works if those functions are never used.

Fixes #171
